+++ mitra_adapters/src/payments/mod.rs
@@ -1,2 +1,3 @@
+pub mod common;
 pub mod monero;
 pub mod subscriptions;